pyproject.toml
[tool.poetry]
name = "electronics-shop"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.1.0"
django = "^5.1.6"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
ipython = "^8.32.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
